% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addAsterisk.R
\name{addAsterisks}
\alias{addAsterisks}
\title{Add Asterisk}
\usage{
addAsterisks(var1, var2, ylim, coords = NULL, diff = Inf, pos = "b",
  p.value = 0.05, cex.as = 1.5, pos.as = "h", densely = F,
  extraspace = 0, ggplot = NULL, facet.vars = NULL, facet.names = NULL)
}
\arguments{
\item{var1}{y-Value}

\item{var2}{x-Value}

\item{ylim}{numeric vector of length 2, giving the y coordinates ranges.}

\item{coords}{x-Achsenticks}

\item{diff}{numeric vector of length 1, giving the maximum length of the brackets}

\item{pos}{1 character code, giving the position of the brackets: "t" top, "b" bottom}

\item{p.value}{numeric vector of length 1, giving the minimum p.value}

\item{cex.as}{asterisk expansion}

\item{pos.as}{1 character code, giving the position of the asterisks: "h" between, "o" outer}

\item{densely}{boolean, try to calculate the densest arrangement}

\item{extraspace}{extra space for the asteriks}

\item{ggplot}{a plot created with ggplot}

\item{facet.vars}{Lay out panels in a grid (vars).}

\item{facet.names}{Lay out panels in a grid (names).}
}
\description{
Adds brackets with asterisks indicating significance levels to boxplots.
}
\examples{
\dontrun{
# Boxplot
boxplot(count ~ spray, data = InsectSprays, ylim = c(0,36))
addAsterisks(InsectSprays$count, InsectSprays$spray, ylim = c(0,36), pos = "t")
# Barplot
counts <- table(mtcars$vs, mtcars$gear)
r <- barplot(counts, main="Car Distribution by Gears and VS",
        xlab="Number of Gears", col=c("darkblue","red"),
        legend = rownames(counts), ylim = c(0,16))
addAsterisks(mtcars$vs, mtcars$gear, ylim = c(0,16),
             coords = r, pos = "t")
# Barplot - beside
counts <- table(mtcars$vs, mtcars$gear)
r <- barplot(counts, main="Car Distribution by Gears and VS",
             xlab="Number of Gears", col=c("darkblue","red"),
             legend = rownames(counts), ylim = c(0,17), beside = T)
addAsterisks(mtcars$vs, mtcars$gear, ylim = c(0,16), coords = r, pos = "t")
}
}

