% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcAsterisk.R
\name{calcAsterisk}
\alias{calcAsterisk}
\title{Calc Asterisk}
\usage{
calcAsterisk(var1, var2, ylim, coords = NULL, diff = Inf, pos = "b",
  p.value = 0.05, pos.as = "h", densely = F, extraspace = 0)
}
\arguments{
\item{var1}{y-Value}

\item{var2}{x-Value}

\item{ylim}{numeric vector of length 2, giving the y coordinates ranges.}

\item{coords}{x-Achsenticks}

\item{diff}{numeric vector of length 1, giving the maximum length of the brackets}

\item{pos}{1 character code, giving the position of the brackets: "t" top, "b" bottom}

\item{p.value}{numeric vector of length 1, giving the minimum p.value}

\item{pos.as}{1 character code, giving the position of the asterisks: "h" between, "o" outer}

\item{densely}{boolean, try to calculate the densest arrangement}
}
\description{
Calculates coords. .
}

