% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/complexArrange.R
\name{complexArrange}
\alias{complexArrange}
\title{Complex Arrange}
\usage{
complexArrange(xx, ylim, factor)
}
\arguments{
\item{xx}{positions of the brackets}

\item{ylim}{numeric vector of length 2, giving the y coordinates ranges.}

\item{factor}{+1,-1}
}
\description{
Calculates the densest arrangement of the brackets.
}

