\name{deg2rad}
\alias{deg2rad}
\title{Converts an angle in degrees to radians
}
\description{
This function converts an angle in degrees to radians.
}
\usage{
deg2rad(degrees)
}
\arguments{
  \item{degrees}{Value of the angle in degrees.
  }
}
\value{
The corresponding value of the angle in radians.
}
\examples{
deg2rad(180)
}
