% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permnArrange.R
\docType{data}
\name{permnArrange}
\alias{permnArrange}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\keyword{created}
\keyword{data}
\keyword{permArrange}
\keyword{with}

