\name{rad2deg}
\alias{rad2deg}
\title{Converts an angle in radians to degrees
}
\description{
This function converts an angle in radians to degrees.
}
\usage{
rad2deg(radians)
}
\arguments{
  \item{radians}{Value of the angle in radians.
  }
}
\value{
The corresponding value of the angle in degrees.
}
\examples{
rad2deg(pi)
}
