% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simpleArrange.R
\name{simpleArrange}
\alias{simpleArrange}
\title{Simple Arrange}
\usage{
simpleArrange(xx, dim, ylim, pos)
}
\arguments{
\item{xx}{positions of the brackets}

\item{dim}{distance > 1}

\item{ylim}{numeric vector of length 2, giving the y coordinates ranges.}

\item{pos}{t == top, b == bottom}
}
\description{
Calculates a simple arrangement of the brackets.
}

